+++ jest.config.js
@@ -8,4 +8,3 @@ module.exports = {
     path.join(__dirname, './test/jest/setupFiles.js'),
   ],
 };
-