+++ test/jest/__mock__/stripesComponents.mock.js
@@ -46,16 +46,15 @@ jest.mock('@folio/stripes/components', () => ({
   // is there a better way to throw it away? If we don't destructure and
   // instead access props.label and props.children, then we get a test
   // failure that the modal isn't visible. oy, dismissible.
-  Modal: jest.fn(({ children, label, dismissible, closeOnBackgroundClick, ...rest }) => {
+  Modal: jest.fn(({ children, label, dismissible, footer, ...rest }) => {
     return (
-      
       <div
         data-test={dismissible ? '' : ''}
-        onClick={closeOnBackgroundClick ? jest.fn() : undefined}
         {...rest}
       >
         <h1>{label}</h1>
         {children}
+        {footer}
       </div>
     );
   }),