+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 
 ## 1.0.0
 
-* New app created with stripes-cli
+* New app created with stripes-cli. Updated module after created with stripes-cli. Refs UIREQMED-1.